% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_occurence.R
\name{calc_p}
\alias{calc_p}
\title{Calculate Probability of Occurance}
\usage{
calc_p(df)
}
\arguments{
\item{df}{}
}
\value{
dataframe of probabilities with upper and lower CI
}
\description{
Calculate Probability of Occurance
}
\examples{

df <- data.frame(A = 1:20, pres = sample(c(TRUE,FALSE), 20, TRUE))
calc_p(df = df)

}
