% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template.R
\name{create_template}
\alias{create_template}
\title{#' Create an ECSA bookdown template for a species of interest.}
\usage{
create_template(stock_name, output_dir = here::here("docs"),
  render_book = T, overwrite = F)
}
\arguments{
\item{stock_name}{A common species name as defined by ??? (eg. atlantic-menhaden, smooth-dogfish)}

\item{output_dir}{The output directory for the compiled bookdown HTML document and supporting files. Default "docs" folder in current working directory}

\item{render_book}{Logical. If TRUE, the generated template will render a bookdown HTML document into the chosen directory and open in a web browser.
If FALSE, the template .Rmd file will be generated in the directory output_dir/stock-name folder}

\item{overwrite}{Logical. If TRUE, output will overwrite any existing template for chosen species.}
}
\value{
A .Rmd file and/or bookdown html document populated with figures
}
\description{
#' Create an ECSA bookdown template for a species of interest.
}
\examples{

create_template(stock_name = "smooth-dogfish", overwrite = T, output_dir = getwd(), render_book = T)

}
