% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_to_strata.R
\name{crop_to_strata}
\alias{crop_to_strata}
\title{A function to mask ecosystem RasterStacks with stock-specific strata of interest. Outputs \code{RasterStacks} as \code{stars} objects for further aggregation.}
\usage{
crop_to_strata(r, stock_name, stock_season, common_name)
}
\arguments{
\item{r}{Filename of an .rdata file containing a RasterStack of ecosystem data to be masked.}

\item{stock_name}{Name of stock to be queried from 'data/stock_data/stock_list.csv'.}

\item{stock_season}{Seasonal designation for strata. Can be \code{"fall"}, \code{"spring"}, or \code{"both"}. However, all seasons do not apply to all stocks.}

\item{common_name}{Common name of species of interest.}
}
\value{

}
\description{
A function to mask ecosystem RasterStacks with stock-specific strata of interest. Outputs \code{RasterStacks} as \code{stars} objects for further aggregation.
}
