% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strata.R
\name{get_strata}
\alias{get_strata}
\title{Download strata from nefsc ftp}
\usage{
get_strata(overwrite = FALSE)
}
\arguments{
\item{overwrite}{}
}
\description{
Download strata from nefsc ftp
}
\examples{

\dontrun{
get_strata(overwrite = FALSE)
}

}
