% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_summary.R
\name{gls_summary}
\alias{gls_summary}
\title{Summarizes GLS fits}
\usage{
gls_summary(mod)
}
\arguments{
\item{mod}{object from nlme fit}
}
\description{
Uses output object from nlme and customizes summary stats
}
