% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRdata.R
\name{loadRData}
\alias{loadRData}
\title{An internal function to load and assign .Rdata files custom names.
Taken from this SO comment: https://stackoverflow.com/a/25455968/9908798}
\usage{
loadRData(fileName)
}
\description{
An internal function to load and assign .Rdata files custom names.
Taken from this SO comment: https://stackoverflow.com/a/25455968/9908798
}
