% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_strata.R
\name{map_strata}
\alias{map_strata}
\title{Map strata}
\usage{
map_strata(common_name, spring_strata, fall_strata, overwrite = FALSE,
  save_plot)
}
\arguments{
\item{common_name}{}

\item{spring_strata}{}

\item{fall_strata}{}

\item{overwrite}{}

\item{save_plot}{}
}
\value{
a ggplot2 object
}
\description{
Map strata
}
\examples{


}
