% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sae.R
\name{plot_sae}
\alias{plot_sae}
\title{Swept area estimate plotting functions}
\usage{
plot_sae(data_season, biomass_, svspp_, k, fill = NA,
  exclude_treatments = NA, include_legend = T, legend_title)
}
\arguments{
\item{data_season}{Either \code{"fall"} or \code{"spring"}. Refers to season of SAE model outputs.}

\item{biomass_}{Either \code{"yes"} or \code{"no"}. If \code{"yes"}, then biomass data (in log10 kg) are chosen for plotting. If \code{"no"}, then abundance data (in log10 N) are chosen instead.}

\item{svspp_}{Survdat species identification code.}

\item{k}{Integer width of rolling-mean window.}

\item{fill}{Fills gaps left by NAs in rolling-mean window. If \code{fill = "extend"}, then the missing values to the left and right of the existing data are
filled with the nearest existing data point, and the middle values are linearly interpolated. Can also be \code{NA}.}

\item{exclude_treatments}{A regex that specifies which \code{treatment} rows should be excluded from the figure output.}

\item{include_legend}{Logical. If \code{TRUE}, then a legend is included for lines in the \code{Plotly} figure output.}

\item{legend_title}{A character string specifying the legend title.}
}
\value{
A \code{Plotly} object.
}
\description{
This function is under construction.
}
\examples{

fall_numbers_sae <- 
spring_sae <- read.csv(here::here("data-raw/spring_sae_tb_sum.csv"), stringsAsFactors = F)

plot_sae(data_season = "fall",
         exclude_treatments = "NNY",
         biomass_ = "no",
         include_zeros = F,
         svspp_ = 36,
         k = 3,
         fill = "extend",
         include_legend = F)
}
