% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_strat.R
\name{resample_strat}
\alias{resample_strat}
\title{Downsample strata shapefiles.}
\usage{
resample_strat(svspp, season, mask_type)
}
\arguments{
\item{svspp}{Input svspp code for species of interest.}

\item{season}{Either "spring" or "fall".}

\item{mask_type}{Specifies raster masking behavior. Can be one of "nes", "gom", "gbk", "sne", or "unit".
If mask is "unit", then returned time series reflect stock boundaries drawn from depth strata.}
}
\value{
Returns a low resolution raster of a given specified region.
}
\description{
This function downsamples NEFSC Bottom Trawl Survey depth strata shapefiles for use in aggregating
stock specific data.
}
\details{
This function is meant to by used internally by \code{\link{stock_env}}.
}
\examples{
sumflo_spring <- match_strat(svspp = 103, mask_type = "unit", season = "spring")
sumflo_fall <- match_strat(svspp = 103, mask_type = "unit", season = "fall")
}
