% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STARS_v18.R
\name{STARS}
\alias{STARS}
\title{STARS_algorithm}
\usage{
STARS(mat.in = c(), window.L = 5, sub.sample.size = 3,
  huber.wt.val = 3, alpha.level = 0.1)
}
\arguments{
\item{mat.in}{A matrix with the time values in column 1 and data vector in column 2.}

\item{window.L}{STARS window length.}

\item{sub.sample.size}{Sub-sample size.}

\item{huber.wt.val}{Huber weight.}

\item{alpha.level}{Alpha level.}
}
\value{

}
\description{
Describe
}
