% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars_to_series.R
\name{stars_to_series}
\alias{stars_to_series}
\title{A function to aggregate \code{stars} objects into derived time series.}
\usage{
stars_to_series(r, stock_name, common_name, stock_season, measure_name,
  data_season, process_to_season = NULL, group_regex = NULL)
}
\arguments{
\item{r}{Filename of an .rdata file containing a RasterStack of ecosystem data to be masked.}

\item{stock_name}{Name of stock to be queried from 'data/stock_data/stock_list.csv'.}

\item{common_name}{Common name of species of interest.}

\item{stock_season}{Seasonal designation for strata. Can be \code{"fall"}, \code{"spring"}, or \code{"both"}. However, all seasons do not apply to all stocks.}

\item{measure_name}{Sets name of grouping variable for summarizing time series.}

\item{data_season}{Seasonal designation for the loaded data set \code{r}.}

\item{process_to_season}{Logical. If \code{TRUE} and data layers within \code{r} are listed monthly, then months are aggregated into fall (Sep., Oct., Nov) and spring (Mar., Apr., May) 
seasons before deriving a mean time series.}

\item{group_regex}{A regex to match a grouping variable within the filename \code{r}. This grouping variable is added as a column in the output \code{data.frame}.}
}
\value{
A \code{data.frame} object.
}
\description{
A function to aggregate \code{stars} objects into derived time series.
}
