% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_plotly.R
\name{tab_plotly}
\alias{tab_plotly}
\title{A function to create standard \code{Plotly} figures for ECSA documents.}
\usage{
tab_plotly(df, showlegend = T, series.name = NULL, add_smoother = F)
}
\arguments{
\item{df}{A wide-format \code{data.frame}, with a Time variable as the first column.}

\item{showlegend}{Logical. If \code{FALSE}, no legend will appear on the figure.}

\item{series.name}{Assigns a name to display with the data series of interest.}

\item{add_smoother}{Adds lines for columns whose names are appended with the string "smooth".}
}
\value{

}
\description{
A function to create standard \code{Plotly} figures for ECSA documents.
}
